Changes errors structure position
