+++ lib/site_mercado.rb
@@ -21,10 +21,10 @@ require 'site_mercado/resources/v1/event'
 require 'site_mercado/resources/v1/product'
 require 'site_mercado/resources/v1/order'
 
-require 'site_mercado/resources/v1/errors/bad_request_error'
-require 'site_mercado/resources/v1/errors/unautorized_error'
-require 'site_mercado/resources/v1/errors/unknown_status_error'
-require 'site_mercado/resources/v1/errors/precondition_failed_error'
+require 'site_mercado/errors/bad_request_error'
+require 'site_mercado/errors/unautorized_error'
+require 'site_mercado/errors/unknown_status_error'
+require 'site_mercado/errors/precondition_failed_error'
 
 require 'site_mercado/helpers/order_parser'
